compressed Get, Post, Update and Delete services into CustomerService
