+++ dot-net-example/Services/Classes/ServiceExtensions.cs
@@ -5,24 +5,9 @@ namespace dot_net_example.Services.Classes
 {
     public static class ServiceExtensions
     {
-        public static IServiceCollection AddGetService(this IServiceCollection services)
+        public static IServiceCollection AddCustomerService(this IServiceCollection services)
         {
-            return services.AddScoped<IGetService, GetService>();
-        }
-
-        public static IServiceCollection AddPostService(this IServiceCollection services)
-        {
-            return services.AddScoped<IPostService, PostService>();
-        }
-
-        public static IServiceCollection AddDeleteService(this IServiceCollection services)
-        {
-            return services.AddScoped<IDeleteService, DeleteService>();
-        }
-
-        public static IServiceCollection AddUpdateService(this IServiceCollection services)
-        {
-            return services.AddScoped<IUpdateService, UpdateService>();
+            return services.AddScoped<ICustomerService, CustomerService>();
         }
     }
 }
