+++ dot-net-example/Services/Interfaces/ICustomerService.cs
@@ -3,10 +3,16 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace dot_net_example.Services.Interfaces
 {
-    public interface IGetService
+    public interface ICustomerService
     {
+        public Task<bool> DeleteCustomer(long id);
+
         public Task<ActionResult<IEnumerable<Customer>>> GetCustomers();
 
         public Task<ActionResult<Customer>> GetCustomer(long id);
+
+        public Task<bool> PostCustomer(Customer customer);
+
+        public Task<bool> PutCustomer(long id, Customer customer);
     }
 }
