--- dot-net-example/Services/Interfaces/IDeleteService.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-
-namespace dot_net_example.Services.Interfaces
-{
-    public interface IDeleteService
-    {
-        public Task<bool> DeleteCustomer(long id);
-    }
-}
