--- dot-net-example/Services/Interfaces/IPostService.cs
@@ -1,10 +1,0 @@
-﻿using dot_net_example.Models;
-using Microsoft.AspNetCore.Mvc;
-
-namespace dot_net_example.Services.Interfaces
-{
-    public interface IPostService
-    {
-        public Task<bool> PostCustomer(Customer customer);
-    }
-}
